Moves pages code from <pagename>.tsx to index.tsx changes the folders' names to reflex URL paths
